+++ sink-connector-lightweight/tests/integration/regression_manual.py
@@ -13,7 +13,7 @@ from integration.helpers.argparser import argparser
 from integration.helpers.common import check_clickhouse_version
 from integration.helpers.common import create_cluster
 from integration.requirements.requirements import *
-from integration.tests.steps.steps_global import *
+from integration.tests.steps.clickhouse import *
 
 xfails = {
     "schema changes/table recreation with different datatypes": [
