--- sink-connector-lightweight/tests/integration/tests/manual_scripts/kafka-list-topics.sh
@@ -1,3 +1,0 @@
-#!/bin/bash
-
-kafkacat -b 127.0.0.1:19092 -L
