--- sink-connector-lightweight/tests/integration/tests/manual_scripts/run_mysqlsnap_load_tests.sh
@@ -1,4 +1,0 @@
-#!/bin/bash
-
-mysql_master_ip_address=`docker inspect -f '{{range.NetworkSettings.Networks}}{{.IPAddress}}{{end}}' mysql-master`
-mysqlslap -h $mysql_master_ip_address -u root -proot --create-schema=test --auto-generate-sql --concurrency=1000 --iterations=1000 --auto-generate-sql-load-type=insert
