+++ sink-connector-lightweight/tests/integration/tests/steps/service_settings.py
